+++ app/controllers/application_controller.rb
@@ -3,4 +3,15 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
   protect_from_forgery with: :null_session, :if => Proc.new { |c| c.request.format == 'application/json' }
+
+  def authorize!
+    if user = User.find_by(access_token: params[:access_token])
+      @current_user = user
+    else
+      message = "#{request.remote_ip} is trying to authorize "
+      message += (params[:access_token] and not params[:access_token].empty?) ? "with invalid access_token #{params[:access_token]}" : "without access_token"
+      logger.error("ProfileController#authorize!") { message }
+      render json: { status: 401, error_msg: 'Access denied: access_token is empty or invalid' }, status: 401
+    end
+  end
 end